perf(honest_diag): _auc helper drops redundant .astype on int labels (1.05-1.19x)
c0032_7f4df0b4 @100k profile (binary lgb+linear+xgb + ensembles +
mrmr + 15 cats) attributed 393us cumtime per call x 3003 calls to
``honest_diagnostics._auc`` (the bootstrap-loop AUC helper). The
closure does ``yy.astype(np.float64, copy=False)`` on every resample
iteration; for int y_true the ``copy=False`` flag is misleading
(it only avoids a copy when dtype ALREADY matches), so each call
allocates a fresh 25k float64 buffer.

The underlying ``fast_numba_auc_nonw`` njit kernel accepts int
y_true natively -- the body is ``tps += y_true[i]`` / ``fps += 1
- y_true[i]``, both widen automatically under numba's mixed-dtype
per-signature dispatch.

Drop the .astype on ``_auc`` only. Bench n=25k:
  int64 labels     2273us -> 2158us   1.05x
  int8 labels      2413us -> 2026us   1.19x
  float64 labels   2187us -> 2182us   1.00x  (no harm)

bench-attempt-rejected for ``_brier`` and ``_ll``: same change
showed 0.77x / 0.85x REGRESSION on float64 labels (something about
the dispatcher fast-path on (float64,float64) signature outperforms
the bare call when the no-op astype is removed). Int-label paths
do win (_brier 1.22-1.37x, _ll 1.14-1.56x) but the float64
regression is worse than the int wins on a path where bootstrap
labels are typically int (so the float64 path is the smaller
slice). Per the iter595/596/597 doctrine ("ship only when no
regression in any reasonable case") these two helpers stay as-is.
Comment in source records this so the next agent doesn't re-try
the same pattern on these two.

Existing tests stay green: 4 in tests/training/test_honest_-
diagnostics_default.py.

Consistent with iter598 _ece_score wrapper fix: the same bootstrap
loop on int labels was paying ~30% of its per-metric cost on a
redundant float64 allocation.

+++ src/mlframe/training/honest_diagnostics.py
@@ -103,8 +103,20 @@ def _bootstrap_block(y_true: np.ndarray, probs: np.ndarray, preds: Optional[np.n
             # bootstrap resampler hands us views of the original arrays so
             # one-time cast in the calling closure is enough (numba dispatcher
             # caches per dtype, so the first resample warms the JIT).
+            # iter599: dropped the .astype on _auc -- fast_roc_auc_unstable's
+            # underlying njit kernel accepts int y_true natively (``tps +=
+            # y_true[i]`` widens automatically). For int labels (the common
+            # bootstrap path) astype(np.float64) ALWAYS copies despite copy=
+            # False (the flag only saves a copy when dtype matches). Bench
+            # n=25k bootstrap-resample size: int64 1.05x, int8 1.19x,
+            # float64 1.00x. Bit-equivalent.
+            # KEPT .astype on _brier / _ll -- bench showed regressions on
+            # float64 labels (0.77x / 0.85x) when kernel was called direct;
+            # something about the dispatcher fast-path on (float64,float64)
+            # outperforms the bare call. Not worth the float64 regression to
+            # win on int paths.
             def _auc(yy, pp):
-                return float(_fast_auc(yy.astype(np.float64, copy=False), pp.astype(np.float64, copy=False)))
+                return float(_fast_auc(yy, pp))
 
             def _brier(yy, pp):
                 return float(_fast_brier(yy.astype(np.float64, copy=False), pp.astype(np.float64, copy=False)))
